# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "description": "Request schema for creating a new API key.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional key description",
            "title": "Description"
          },
          "name": {
            "description": "Human-readable key name",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "APIKeyCreate",
        "type": "object"
      },
      "APIKeyCreated": {
        "description": "Response schema when a key is first created (includes the actual key).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "warning": {
            "default": "\u26a0\ufe0f Save this key now - it won't be shown again!",
            "title": "Warning",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "key",
          "key_prefix",
          "created_at",
          "created_by"
        ],
        "title": "APIKeyCreated",
        "type": "object"
      },
      "APIKeyList": {
        "description": "Response schema for listing API keys.",
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/APIKeyResponse"
            },
            "title": "Keys",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "keys",
          "total"
        ],
        "title": "APIKeyList",
        "type": "object"
      },
      "APIKeyResponse": {
        "description": "Response schema for API key (without the actual key).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "key_prefix",
          "created_at",
          "last_used_at",
          "revoked_at",
          "is_active",
          "created_by"
        ],
        "title": "APIKeyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Docker vulnerability scanner and remediation dashboard powered by Trivy",
    "title": "VulnForge",
    "version": "4.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/api-keys": {
      "get": {
        "description": "List all API keys.\n\nBy default, only active (non-revoked) keys are returned.\n\nArgs:\n    include_revoked: Include revoked keys in results\n    db: Database session\n\nReturns:\n    List of API keys (without the actual key values)",
        "operationId": "list_api_keys_api_v1_api_keys_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_revoked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Revoked",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Api Keys",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "Create a new API key.\n\nReturns the full API key in the response - this is the ONLY time it will be shown.\nThe key should be saved immediately by the client.\n\nArgs:\n    key_data: Name and optional description for the key\n    db: Database session\n\nReturns:\n    APIKeyCreated with full key (shown only once)",
        "operationId": "create_api_key_api_v1_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Api Key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "delete": {
        "description": "Revoke an API key (soft delete).\n\nThe key will no longer be valid for authentication but remains in the database\nfor audit purposes.\n\nArgs:\n    key_id: ID of the API key to revoke\n    db: Database session\n\nReturns:\n    Updated API key with revoked_at timestamp",
        "operationId": "revoke_api_key_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Api Key",
        "tags": [
          "api-keys"
        ]
      },
      "get": {
        "description": "Get details of a specific API key.\n\nArgs:\n    key_id: ID of the API key\n    db: Database session\n\nReturns:\n    API key details (without the actual key value)",
        "operationId": "get_api_key_api_v1_api_keys__key_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Api Key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  }
}